services/extract-faces-service/requirements.txt
fastapi==0.115.5
uvicorn==0.34.0
pydantic==2.10.3
python-dotenv==1.1.1
structlog==25.4.0
requests==2.32.5
python-multipart==0.0.20
deepface==0.0.93
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==11.0.0
tf-keras==2.18.0
